@@ -1,6 +1,6 @@
 module github.com/calqs/gopkg/crypt
 
-go 1.24.6
+go 1.25.0
 
 require golang.org/x/crypto v0.41.0
 
